move mod install related functions to a separate file in prep for adding more mod init functions
